+++ src/main/resources/com.cafein.mapper/UserMapper.xml
@@ -3,6 +3,6 @@
   PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.cafein.mapper.UserMapper">
+<mapper namespace="com.cafein.user.UserMapper">
 
  </mapper>
